fix(Builder): remove some methods that triggers error for dynamic property declaration (#6)
* fix(Builder): remove some methods that triggers error for dynamic property declaration

A long as this methods should not be called and there are other methods that
are equivalent and they are the ones that should be used, this change won't be
considered a BC

* docs(examples/example-1.php): add long text as attribute on usage examples

Works better than gTrace!! gTrace has a native restriction on attribute length, 256 is the max length for them. Jaegger or Zipkin has not.